(#1326) rename the jetstream adapter to choria_streams
